requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic==2.6.3
beautifulsoup4
openai
python-multipart